Moved useGalleryParams to galleryTreeContext
