+++ src/hooks/usePathSegments.ts
@@ -1,7 +1,6 @@
 import { useMemo } from 'react';
 import { useGalleryTreeContext } from '@/contexts/galleryTree';
 import { useNavigationToolsContext } from '@/contexts/navigationTools/NavigationToolsProvider';
-import { useGalleryParams } from '@/hooks/useGalleryParams';
 import { PathMap } from '@/types/galleryTreeContext';
 import type { TreeImageGroup } from '@/types/ImageGroup';
 
@@ -15,8 +14,7 @@ export interface UsePathSegments {
 }
 
 export const usePathSegments = (): UsePathSegments => {
-  const { path: currentPath, getUrl } = useGalleryParams();
-  const { root, paths } = useGalleryTreeContext();
+  const { path: currentPath, getUrl, root, paths } = useGalleryTreeContext();
   const { getImagePageIndexInGroup, navigateToGroup } = useNavigationToolsContext();
 
   const segments = useMemo<Segment[]>(() => {
